Refactor the socks5 code into multiple files
